Move initialisation logic into it's own function
